Created new store for chat Improved gallery (size, switch and close with keyboard, video support) Refactor
