+++ components/CtmModal/CtmModalGallery/index.vue
@@ -1,41 +1,51 @@
 <template>
-  <div class="ctm-modal">
-    <div class="gallery-modal">
-      <div class="gallery-modal__container">
-        <div class="gallery-modal__pager">
-          <img
-            :src="getUrl"
-            alt=""
-            class="gallery-modal__image"
-          >
-          <div class="gallery-modal__close">
-            <div
-              class="control-btn control-btn_bg"
-              @click="hide"
-            >
-              <span class="icon-close_big" />
-            </div>
-          </div>
+  <div
+    ref="GalleryModal"
+    class="ctm-modal__box gallery-modal"
+  >
+    <div class="gallery-modal__container">
+      <div
+        class="gallery-modal__pager"
+      >
+        <video
+          v-if="isVideo"
+          :src="getUrl"
+          controls
+          class="gallery-modal__image"
+        />
+        <img
+          v-else
+          :src="getUrl"
+          alt=""
+          class="gallery-modal__image"
+        >
+        <div class="gallery-modal__close">
           <div
-            v-if="options.files.length > 1"
-            class="gallery-modal__controls"
+            class="control-btn control-btn_bg"
+            @click="hide"
           >
-            <div class="gallery-modal__switches">
-              <div
-                class="control-btn"
-                @click="turnOver(false)"
-              >
-                <span class="icon-short_left" />
-              </div>
-              <div class="gallery-modal__title">
-                {{ $tc('gallery.counter', currIndex >= 0 ? currIndex + 1 : options.index + 1) + options.files.length }}
-              </div>
-              <div
-                class="control-btn"
-                @click="turnOver(true)"
-              >
-                <span class="icon-short_right" />
-              </div>
+            <span class="icon-close_big" />
+          </div>
+        </div>
+        <div
+          v-if="options.files.length > 1"
+          class="gallery-modal__controls"
+        >
+          <div class="gallery-modal__switches">
+            <div
+              class="control-btn"
+              @click="turnOver(false)"
+            >
+              <span class="icon-short_left" />
+            </div>
+            <div class="gallery-modal__title">
+              {{ $tc('gallery.counter', currIndex >= 0 ? currIndex + 1 : options.index + 1) + options.files.length }}
+            </div>
+            <div
+              class="control-btn"
+              @click="turnOver(true)"
+            >
+              <span class="icon-short_right" />
             </div>
           </div>
         </div>
@@ -58,12 +68,40 @@ export default {
     ...mapGetters({
       options: 'modals/getOptions',
     }),
+    isVideo() {
+      const file = this.getCurrFile();
+      return file.type === 'video';
+    },
     getUrl() {
-      const { options: { files, index }, currIndex } = this;
-      return files[currIndex >= 0 ? currIndex : index].url;
+      const file = this.getCurrFile();
+      return file.url;
     },
   },
+  mounted() {
+    document.addEventListener('keydown', (ev) => this.handleClickBtn(ev));
+  },
+  destroyed() {
+    document.removeEventListener('keydown', (ev) => this.handleClickBtn(ev));
+  },
   methods: {
+    handleClickBtn(ev) {
+      if (this.options.files.length > 1) {
+        if (ev.keyCode === 39) {
+          this.turnOver(true);
+          return;
+        }
+        if (ev.keyCode === 37) {
+          this.turnOver(false);
+          return;
+        }
+      }
+
+      if (ev.keyCode === 27) this.hide();
+    },
+    getCurrFile() {
+      const { options: { files, index }, currIndex } = this;
+      return files[currIndex >= 0 ? currIndex : index];
+    },
     turnOver(isForward) {
       const { options: { files, index }, currIndex } = this;
       const trackedIndex = currIndex >= 0 ? currIndex : index;
@@ -81,29 +119,26 @@ export default {
 
 <style lang="scss" scoped>
 .gallery-modal {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-
-  &__container {
-    max-width: 1180px;
-  }
+  width: 1180px !important;
+  height: 80vh !important;
+  max-width: unset !important;
 
   &__pager {
     position: relative;
     box-shadow: 0 85px 147px rgba(10, 27, 61, 0.17), 0 47.1676px 61.4131px rgba(10, 27, 61, 0.078707), 0 26.7219px 32.8344px rgba(10, 27, 61, 0.0629546), 0 14.4955px 18.4067px rgba(10, 27, 61, 0.0598272), 0 6.96225px 9.77565px rgba(10, 27, 61, 0.0584222), 0 2.43911px 4.06787px rgba(10, 27, 61, 0.0492837);
     border-radius: 6px;
-    min-width: 375px;
-    min-height: 300px;
     display: flex;
     align-items: center;
     justify-content: center;
     background-color: #C4C4C4;
+    width: 1180px;
+    height: 80vh;
   }
 
   &__image {
     border-radius: 6px;
+    max-width: 1180px;
+    max-height: 80vh;
   }
 
   &__close {
