+++ store/chat/state.js
@@ -1,0 +1,21 @@
+export default () => ({
+  chats: {
+    list: [],
+    count: 0,
+  },
+  messages: {
+    list: [],
+    count: 0,
+    chat: null,
+  },
+  groupChatUsers: {
+    count: 0,
+    list: [],
+  },
+  messagesFilter: {
+    topOffset: 0,
+    bottomOffset: 0,
+    canLoadToBottom: false,
+    canLoadToTop: false,
+  },
+});
