+++ store/data/getters.js
@@ -11,7 +11,6 @@ export default {
   getBadgeList: (state) => state.badgeList,
   getSocials: (state) => state.socials,
   getDisputes: (state) => state.disputes,
-  getMessages: (state) => state.messages,
   getTransactions: (state) => state.transactions,
   getTransactionsData: (state) => state.transactionsData,
   getRespondedList: (state) => state.respondedList,
@@ -24,11 +23,5 @@ export default {
   getEducations: (state) => state.userInfo?.educations,
   getWorkExp: (state) => state.userInfo?.workExp,
   getUserDesc: (state) => state.userInfo?.desc,
-  getChats: (state) => state.chats,
   notificationsConnectionStatus: (state) => state.connections,
-  getCurrChatId: (state) => (state.messages.chat?.id || ''),
-  getLastMessageId: (state) => (!state.messagesFilter.canLoadToBottom && state.messages.list.length ? state.messages.list[state.messages.list.length - 1].id : null),
-  getMessagesFilter: (state) => state.messagesFilter,
-  getGroupChatUsers: (state) => state.groupChatUsers,
-  getChatMembers: (state) => (state.messages.chat?.members || []),
 };
