+++ store/modals/mutations.js
@@ -17,4 +17,9 @@ export default {
   setIsShowWelcome(state, value) {
     state.showWelcomeModal = value;
   },
+  setCurrOptionByKey(state, optionsArr) {
+    optionsArr.forEach(({ key, val }) => {
+      state.options[key] = val;
+    });
+  },
 };
